fix access to db model in Thymeleaf templates
fixes #434
