+++ jbake-core/src/test/resources/fixture/groovyMarkupTemplates/index.tpl
@@ -5,11 +5,12 @@ layout 'layout/main.tpl',
 
                 div(class:"row-fluid marketing"){
                     div(class:"span12"){
-                        published_posts[0..<2].each { post ->
+                        published_posts.each { post ->
                             h4 { a(href:"${post.uri}","${post.title}") }
                             p("${post.date.format("dd MMMM yyyy")} - ${post.body.substring(0, 150)}...")
                         }
                         a(href:"/archive.html",'Archive')
                     }
                 }
+            span("${db.getPublishedPages().size()}")
         }
